requirements.txt
argcomplete==1.9.3
base58==1.0.0
boto3==1.7.62
botocore==1.10.62
certifi==2018.4.16
cfn-flip==1.0.3
chardet==3.0.4
click==6.7
docutils==0.14
durationpy==0.5
Flask==1.0.2
Flask-PyMongo==2.0.1
future==0.16.0
hjson==3.0.1
idna==2.7
itsdangerous==0.24
Jinja2==2.10
jmespath==0.9.3
kappa==0.6.0
lambda-packages==0.20.0
MarkupSafe==1.0
placebo==0.8.1
pyee==5.0.0
pymongo==3.7.1
python-dateutil==2.6.1
python-dotenv==0.8.2
python-slugify==1.2.4
PyYAML==3.12
requests==2.19.1
s3transfer==0.1.13
six==1.11.0
slackclient==1.2.1
slackeventsapi==1.1.0
toml==0.9.4
tqdm==4.19.1
troposphere==2.3.1
Unidecode==1.0.22
urllib3==1.23
websocket-client==0.48.0
Werkzeug==0.14.1
wsgi-request-logger==0.4.6
zappa==0.46.2
